implement reporting
